fix: YouTubeとWowzaで再生時間が取得されない
視聴ログ関連ではないイベントハンドリングをPlayerTrackerから
分離したことに起因
